Refactor x2
