+++ client/src/components/deleteItem.helper.ts
