+++ client/src/features/category/CategoryList.tsx
@@ -3,7 +3,7 @@ import { useMakeRequest } from "../../networking/network";
 import { Category } from "./category";
 import { Grid, Paper, Typography } from "@material-ui/core";
 import { DeleteItem } from "../../components/DeleteItem";
-import { deleteRequest } from "../../components/DeleteItem.h";
+import { deleteRequest } from "../../components/deleteItem.helper";
 
 export const CategoryList: React.FC = () => {
   const [selectedCategory, setSelectedCategory] = useState<Category>();
